+++ src/ir/effects.h
@@ -28,8 +28,8 @@ namespace wasm {
 class EffectAnalyzer {
 public:
   EffectAnalyzer(const PassOptions& passOptions, Module& module)
-    : ignoreImplicitTraps(passOptions.ignoreImplicitTraps),
-      trapsNeverHappen(passOptions.trapsNeverHappen),
+    : trapsNeverHappen(passOptions.trapsNeverHappen),
+      targetJS(passOptions.targetJS),
       funcEffectsMap(passOptions.funcEffectsMap), module(module),
       features(module.features) {}
 
@@ -45,8 +45,8 @@ public:
     walk(func);
   }
 
-  bool ignoreImplicitTraps;
   bool trapsNeverHappen;
+  bool targetJS;
   std::shared_ptr<FuncEffectsMap> funcEffectsMap;
   Module& module;
   FeatureSet features;
@@ -117,7 +117,7 @@ public:
   bool writesStruct = false;
   bool readsArray = false;
   bool writesArray = false;
-  // A trap, either from an unreachable instruction, or from an implicit trap
+  // A trap, either from an unreachable instruction, or from an XXX trap
   // that we do not ignore (see below).
   //
   // Note that we ignore trap differences, so it is ok to reorder traps with
@@ -135,9 +135,6 @@ public:
   // VMs would trap eventually, and the same for potentially infinite recursion,
   // etc.
   bool trap = false;
-  // A trap from an instruction like a load or div/rem, which may trap on corner
-  // cases. If we do not ignore implicit traps then these are counted as a trap.
-  bool implicitTrap = false;
   // An atomic load/store/RMW/Cmpxchg or an operator that has a defined ordering
   // wrt atomics (e.g. memory.grow)
   bool isAtomic = false;
@@ -359,7 +356,6 @@ public:
     readsArray = readsArray || other.readsArray;
     writesArray = writesArray || other.writesArray;
     trap = trap || other.trap;
-    implicitTrap = implicitTrap || other.implicitTrap;
     trapsNeverHappen = trapsNeverHappen || other.trapsNeverHappen;
     isAtomic = isAtomic || other.isAtomic;
     throws_ = throws_ || other.throws_;
@@ -573,24 +569,28 @@ private:
     void visitLoad(Load* curr) {
       parent.readsMemory = true;
       parent.isAtomic |= curr->isAtomic;
-      parent.implicitTrap = true;
+      if (!parent.targetJS) {
+        parent.trap = true;
+      }
     }
     void visitStore(Store* curr) {
       parent.writesMemory = true;
       parent.isAtomic |= curr->isAtomic;
-      parent.implicitTrap = true;
+      if (!parent.targetJS) {
+        parent.trap = true;
+      }
     }
     void visitAtomicRMW(AtomicRMW* curr) {
       parent.readsMemory = true;
       parent.writesMemory = true;
       parent.isAtomic = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitAtomicCmpxchg(AtomicCmpxchg* curr) {
       parent.readsMemory = true;
       parent.writesMemory = true;
       parent.isAtomic = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitAtomicWait(AtomicWait* curr) {
       parent.readsMemory = true;
@@ -599,7 +599,7 @@ private:
       // of as a write.
       parent.writesMemory = true;
       parent.isAtomic = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitAtomicNotify(AtomicNotify* curr) {
       // AtomicNotify doesn't strictly write memory, but it does modify the
@@ -608,7 +608,7 @@ private:
       parent.readsMemory = true;
       parent.writesMemory = true;
       parent.isAtomic = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitAtomicFence(AtomicFence* curr) {
       // AtomicFence should not be reordered with any memory operations, so we
@@ -624,7 +624,7 @@ private:
     void visitSIMDShift(SIMDShift* curr) {}
     void visitSIMDLoad(SIMDLoad* curr) {
       parent.readsMemory = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitSIMDLoadStoreLane(SIMDLoadStoreLane* curr) {
       if (curr->isLoad()) {
@@ -632,27 +632,27 @@ private:
       } else {
         parent.writesMemory = true;
       }
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitMemoryInit(MemoryInit* curr) {
       parent.writesMemory = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitDataDrop(DataDrop* curr) {
       // data.drop does not actually write memory, but it does alter the size of
       // a segment, which can be noticeable later by memory.init, so we need to
       // mark it as having a global side effect of some kind.
       parent.writesMemory = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitMemoryCopy(MemoryCopy* curr) {
       parent.readsMemory = true;
       parent.writesMemory = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitMemoryFill(MemoryFill* curr) {
       parent.writesMemory = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitConst(Const* curr) {}
     void visitUnary(Unary* curr) {
@@ -665,7 +665,9 @@ private:
         case TruncSFloat64ToInt64:
         case TruncUFloat64ToInt32:
         case TruncUFloat64ToInt64: {
-          parent.implicitTrap = true;
+          if (!parent.targetJS) {
+            parent.trap = true;
+          }
           break;
         }
         default: {
@@ -682,19 +684,21 @@ private:
         case DivUInt64:
         case RemSInt64:
         case RemUInt64: {
-          // div and rem may contain implicit trap only if RHS is
-          // non-constant or constant which equal zero or -1 for
-          // signed divisions. Reminder traps only with zero
-          // divider.
-          if (auto* c = curr->right->dynCast<Const>()) {
-            if (c->value.isZero()) {
-              parent.implicitTrap = true;
-            } else if ((curr->op == DivSInt32 || curr->op == DivSInt64) &&
-                       c->value.getInteger() == -1LL) {
-              parent.implicitTrap = true;
+          if (!parent.targetJS) {
+            // div and rem may contain implicit trap only if RHS is
+            // non-constant or constant which equal zero or -1 for
+            // signed divisions. Reminder traps only with zero
+            // divider.
+            if (auto* c = curr->right->dynCast<Const>()) {
+              if (c->value.isZero()) {
+                parent.trap = true;
+              } else if ((curr->op == DivSInt32 || curr->op == DivSInt64) &&
+                         c->value.getInteger() == -1LL) {
+                parent.trap = true;
+              }
+            } else {
+              parent.trap = true;
             }
-          } else {
-            parent.implicitTrap = true;
           }
           break;
         }
@@ -729,11 +733,11 @@ private:
     void visitRefEq(RefEq* curr) {}
     void visitTableGet(TableGet* curr) {
       parent.readsTable = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitTableSet(TableSet* curr) {
       parent.writesTable = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitTableSize(TableSize* curr) { parent.readsTable = true; }
     void visitTableGrow(TableGrow* curr) {
@@ -745,12 +749,12 @@ private:
     }
     void visitTableFill(TableFill* curr) {
       parent.writesTable = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitTableCopy(TableCopy* curr) {
       parent.readsTable = true;
       parent.writesTable = true;
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitTry(Try* curr) {
       if (curr->delegateTarget.is()) {
@@ -777,7 +781,7 @@ private:
         parent.throws_ = true;
       }
       // traps when the arg is null
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitNop(Nop* curr) {}
     void visitUnreachable(Unreachable* curr) { parent.trap = true; }
@@ -792,7 +796,7 @@ private:
     void visitI31Get(I31Get* curr) {
       // traps when the ref is null
       if (curr->i31->type.isNullable()) {
-        parent.implicitTrap = true;
+        parent.trap = true;
       }
     }
     void visitCallRef(CallRef* curr) {
@@ -808,7 +812,7 @@ private:
       }
       // traps when the call target is null
       if (curr->target->type.isNullable()) {
-        parent.implicitTrap = true;
+        parent.trap = true;
       }
 
       parent.calls = true;
@@ -820,7 +824,7 @@ private:
     void visitRefTest(RefTest* curr) {}
     void visitRefCast(RefCast* curr) {
       // Traps if the ref is not null and the cast fails.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitBrOn(BrOn* curr) { parent.breakTargets.insert(curr->name); }
     void visitStructNew(StructNew* curr) {}
@@ -840,7 +844,7 @@ private:
       }
       // traps when the arg is null
       if (curr->ref->type.isNullable()) {
-        parent.implicitTrap = true;
+        parent.trap = true;
       }
     }
     void visitStructSet(StructSet* curr) {
@@ -851,19 +855,19 @@ private:
       parent.writesStruct = true;
       // traps when the arg is null
       if (curr->ref->type.isNullable()) {
-        parent.implicitTrap = true;
+        parent.trap = true;
       }
     }
     void visitArrayNew(ArrayNew* curr) {}
     void visitArrayNewData(ArrayNewData* curr) {
       // Traps on out of bounds access to segments or access to dropped
       // segments.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitArrayNewElem(ArrayNewElem* curr) {
       // Traps on out of bounds access to segments or access to dropped
       // segments.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitArrayNewFixed(ArrayNewFixed* curr) {}
     void visitArrayGet(ArrayGet* curr) {
@@ -873,7 +877,7 @@ private:
       }
       parent.readsArray = true;
       // traps when the arg is null or the index out of bounds
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitArraySet(ArraySet* curr) {
       if (curr->ref->type.isNull()) {
@@ -882,7 +886,7 @@ private:
       }
       parent.writesArray = true;
       // traps when the arg is null or the index out of bounds
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitArrayLen(ArrayLen* curr) {
       if (curr->ref->type.isNull()) {
@@ -891,7 +895,7 @@ private:
       }
       // traps when the arg is null
       if (curr->ref->type.isNullable()) {
-        parent.implicitTrap = true;
+        parent.trap = true;
       }
     }
     void visitArrayCopy(ArrayCopy* curr) {
@@ -902,7 +906,7 @@ private:
       parent.readsArray = true;
       parent.writesArray = true;
       // traps when a ref is null, or when out of bounds.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitArrayFill(ArrayFill* curr) {
       if (curr->ref->type.isNull()) {
@@ -911,7 +915,7 @@ private:
       }
       parent.writesArray = true;
       // Traps when the destination is null or when out of bounds.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     template<typename ArrayInit> void visitArrayInit(ArrayInit* curr) {
       if (curr->ref->type.isNull()) {
@@ -921,7 +925,7 @@ private:
       parent.writesArray = true;
       // Traps when the destination is null, when out of bounds in source or
       // destination, or when the source segment has been dropped.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitArrayInitData(ArrayInitData* curr) { visitArrayInit(curr); }
     void visitArrayInitElem(ArrayInitElem* curr) { visitArrayInit(curr); }
@@ -931,7 +935,7 @@ private:
         return;
       }
       // traps when the arg is not valid
-      parent.implicitTrap = true;
+      parent.trap = true;
       // Note: We could be more precise here and report the lack of a possible
       // trap if the input is non-nullable (and also of the right kind for
       // RefAsFunc etc.). However, we have optimization passes that will
@@ -943,7 +947,7 @@ private:
     }
     void visitStringNew(StringNew* curr) {
       // traps when ref is null
-      parent.implicitTrap = true;
+      parent.trap = true;
       if (curr->op != StringNewFromCodePoint) {
         parent.readsArray = true;
       }
@@ -951,40 +955,40 @@ private:
     void visitStringConst(StringConst* curr) {}
     void visitStringMeasure(StringMeasure* curr) {
       // traps when ref is null.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitStringEncode(StringEncode* curr) {
       // traps when ref is null or we write out of bounds.
-      parent.implicitTrap = true;
+      parent.trap = true;
       parent.writesArray = true;
     }
     void visitStringConcat(StringConcat* curr) {
       // traps when an input is null.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitStringEq(StringEq* curr) {
       if (curr->op == StringEqCompare) {
         // traps when either input is null.
         if (curr->left->type.isNullable() || curr->right->type.isNullable()) {
-          parent.implicitTrap = true;
+          parent.trap = true;
         }
       }
     }
     void visitStringWTF16Get(StringWTF16Get* curr) {
       // traps when ref is null.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitStringSliceWTF(StringSliceWTF* curr) {
       // traps when ref is null.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitContBind(ContBind* curr) {
       // traps when curr->cont is null ref.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitContNew(ContNew* curr) {
       // traps when curr->func is null ref.
-      parent.implicitTrap = true;
+      parent.trap = true;
     }
     void visitResume(Resume* curr) {
       // This acts as a kitchen sink effect.
@@ -992,7 +996,7 @@ private:
 
       // resume instructions accept nullable continuation references and trap
       // on null.
-      parent.implicitTrap = true;
+      parent.trap = true;
 
       if (parent.features.hasExceptionHandling() && parent.tryDepth == 0) {
         parent.throws_ = true;
@@ -1035,7 +1039,6 @@ public:
     WritesMemory = 1 << 7,
     ReadsTable = 1 << 8,
     WritesTable = 1 << 9,
-    ImplicitTrap = 1 << 10,
     IsAtomic = 1 << 11,
     Throws = 1 << 12,
     DanglingPop = 1 << 13,
@@ -1074,9 +1077,6 @@ public:
     if (writesTable) {
       effects |= SideEffects::WritesTable;
     }
-    if (implicitTrap) {
-      effects |= SideEffects::ImplicitTrap;
-    }
     if (trapsNeverHappen) {
       effects |= SideEffects::TrapsNeverHappen;
     }
@@ -1110,12 +1110,6 @@ public:
 private:
   void post() {
     assert(tryDepth == 0);
-
-    if (ignoreImplicitTraps) {
-      implicitTrap = false;
-    } else if (implicitTrap) {
-      trap = true;
-    }
   }
 };