+++ CHANGELOG.md
@@ -41,6 +41,9 @@ v118
  - The build-time option to use legacy WasmGC opcodes is removed.
  - The strings in `string.const` instructions must now be valid WTF-8.
  - The `TraverseCalls` flag for `ExpressionRunner` is removed.
+ - Remove the "ignore implicit effects" options. That is replaced by the
+   now-mature "traps never happen" option which has proven far more useful.
+   - The C API `BinaryenSideEffectImplicitTrap` is removed.
 
 v117
 ----